Fix 64-bit build on Mac
BUG=http://code.google.com/p/skia/issues/detail?id=524
Thanks to mmaerean@adobe.com for the patch.
Review URL: https://codereview.appspot.com/6022058

git-svn-id: http://skia.googlecode.com/svn/trunk@3686 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ include/core/SkFontHost.h
@@ -17,7 +17,7 @@ class SkDescriptor;
 class SkStream;
 class SkWStream;
 
-typedef uint32_t SkFontTableTag;
+typedef uintptr_t SkFontTableTag;
 
 /** \class SkFontHost
